Enable tests and build scripts to run on Windows
There appears to be a bug in tslint's test harness on Windows where CRLF
line endings get parsed incorrectly, causing multiline errors in the
test suite to fail. Additionally, tslint will fail to parse the test
files at all if the line endings are converted to Unix-style. Until this
is fixed upstream, the solution I found was just to place the multiline
expressions on a single line. Inelegant, unfortunately.

The build scripts now invoke rimraf (which is now an added development
dependency) in order to implement `rm -rf` across platforms. Next
perhaps I can contribute something interesting while I'm stuck on
Windows.

Finally, as an unrelated change, I fixed a typo I noticed in the
no-if-statement rule's error message and added a clarification to it.
The error message now reads, "Unexpected if, use a conditional
expression (ternary operator) instead." I can put that in a separate
pull request, if so desired, but it seemed too minor to matter.